Move `mc-sgx-dcap-types::Error` to a separate module
Previously the `Quote3Error` was inside of the
`mc-sgx-dcap-types::quote3` module. In order to support further
expanding the `Quote3` type the error has been moved to a separate
module `mc-sgx-dcap-types::error`.
